refactor: rename run_env.py -> train_env.py
